Wip: icons sidebar

+++ db/migrate/20250324134708_create_g_tipo_usuarios.rb
@@ -2,17 +2,20 @@
 
 class CreateGTipoUsuarios < ActiveRecord::Migration[7.2]
   def up
-    create_table :g_tipo_usuarios do |t|
-      t.string   :nome, null: false
-      t.text     :descricao
-      t.string   :created_by
-      t.string   :updated_by
-      t.datetime :deleted_at
-      t.timestamps
+    unless table_exists?(:g_tipo_usuarios)
+      create_table :g_tipo_usuarios do |t|
+        t.string   :nome, null: false
+        t.text     :descricao
+        t.string   :created_by
+        t.string   :updated_by
+        t.datetime :deleted_at
+        t.timestamps
+      end
     end
   end
-
+  
   def down
-    drop_table :g_tipo_usuarios
+    drop_table :g_tipo_usuarios if table_exists?(:g_tipo_usuarios)
   end
+  
 end
